+++ Coimbra.Services.Events/EventCallbacks.cs
@@ -13,5 +13,10 @@ namespace Coimbra.Services.Events
         internal static readonly Dictionary<EventHandle, EventContextHandler<TEvent>> Value = new Dictionary<EventHandle, EventContextHandler<TEvent>>(1);
 
         internal static readonly Func<EventHandle, bool> RemoveHandler = Value.Remove;
+
+        internal static readonly Func<EventHandle, List<DelegateListener>, int> GetListenersHandler = delegate(EventHandle handle, List<DelegateListener> list)
+        {
+            return Value.TryGetValue(handle, out EventContextHandler<TEvent> handler) ? handler.GetListeners(list) : 0;
+        };
     }
 }